Designer: KP-7122 fix unit tests

+++ src/Tests/WB.Tests.Integration/InterviewTests/Variables/when_Answering_question_which_changes_value_of_a_variable.cs
@@ -40,7 +40,7 @@ namespace WB.Tests.Integration.InterviewTests.Variables
 
         It should_raise_VariablesValuesChanged_event_for_the_variable = () =>
             eventContext.ShouldContainEvent<VariablesChanged>(@event
-                => (int?) @event.ChangedVariables[0].NewValue == 6 && @event.ChangedVariables[0].Identity.Id== variableId);
+                => (long?) @event.ChangedVariables[0].NewValue == 6 && @event.ChangedVariables[0].Identity.Id== variableId);
 
         private static EventContext eventContext;
         private static Interview interview;
